http codec update
